Angular Modal popup notiifcation Added and scroll message bar

+++ client/client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { MessageService } from '../_services/message.service';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 
 
 @Component({
@@ -19,7 +20,7 @@ export class MessagesComponent implements OnInit{
  //adding a load flag as this allow thing to be ready before loading it
  loading = false;
 
-  constructor(private messageService: MessageService){}
+  constructor(private messageService: MessageService, private confirmService: ConfirmService){}
 
   ngOnInit(): void {
     this.loadMessages();
@@ -38,10 +39,17 @@ export class MessagesComponent implements OnInit{
 
 
   deleteMessage(id: number){
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
-     ;
-    })
+    this.confirmService.confirm('Confirm delete message ', ' this cannot be undone')
+    .subscribe(result => {
+      if(result)
+      {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+         ;
+        })
+      }    
+    });
+    
   }
   pageChanged(event: any){
     if(this.pageNumber !== event.page)
